+++ src/application/mod.rs
@@ -27,6 +27,7 @@ pub mod new_item;
 pub mod notes;
 pub mod pim_command;
 pub mod pop_sync;
+pub mod reading_style;
 pub mod receipts;
 pub mod reminders;
 pub mod reply;